todo.py: Delete endp[oint update- user authentication error raise, user dependency injection,filter(),

+++ Todo_app/routers/todo.py
@@ -95,9 +95,15 @@ async def update_todo(db:db_dependency,
     db.commit()
 
 @router.delete("\todo\{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model=db.query(Todo).filter(Todo.id==todo_id).first()
+async def delete_todo(db:db_dependency,user:user_dependency,
+                      todo_id:int=Path(gt=0)):
+    
+    if user is None:
+        raise HTTPException(status_code=401,detail='authentication failed')
+    
+    todo_model=db.query(Todo).filter(Todo.id==todo_id).filter(Todo.owner_id==user.get('id')).first()
+
     if todo_model is None:    
         raise HTTPException(status_code=404,detail="Todo not found.")
-    db.query(Todo).filter(Todo.id==todo_id).delete()
+    db.query(Todo).filter(Todo.id==todo_id).filter(Todo.owner_id==user.get('id')).delete()
     db.commit()
